Pass value types test

+++ openfisca_uk/microdata/frs/config.py
@@ -748,7 +748,7 @@ class accommodation_type(Variable):
 
 
 class num_bedrooms(Variable):
-    value_type = float
+    value_type = int
     entity = Household
     label = "The number of bedrooms in the house"
     definition_period = YEAR
